Support for ModR/M byte except SIB cases
This adds support for all of the ModR/M byte possibilities, except for
the cases that indicate to use an additional SIB byte, potentially with
a displacement.

This also adds a test suite to check various combinations of
instructions which we now support with the ModR/M byte, with the
register-to-register varieties, and with the single immediate operand.